requirements.txt
Scrapy=2.5.0
Django~=3.2
requests
scrapy-djangoitem~=1.1.1
python-decouple
djangorestframework